fix: Skip preview version .NET SDK only when running tests
